Add buttons for assignee and workflow owner
