+++ d1.workflow.frontend.js
@@ -1,4 +1,3 @@
-import { makePostCall, addNameList, findPos, mapAssignee, mapFromElement} from "./d1.workflow.util.js"
 
 (function(jq$){
 	'use strict';
@@ -38,7 +37,7 @@ import { makePostCall, addNameList, findPos, mapAssignee, mapFromElement} from "
 
         jq$('.user-flexchild').append(
             '<ul id="user-list" class="content-list">'
-            + '<li class="flex-first-row">User</li>'
+            + '<li class="flex-first-row">Name</li>'
             + '</ul>');
 
         jq$('.assigned-flexchild').append(
@@ -54,7 +53,7 @@ import { makePostCall, addNameList, findPos, mapAssignee, mapFromElement} from "
             +'</div>');
 
         //Append list to table for availble users in the current PO/CU
-        addNameList("default");
+        addNameList("default", names.assignee);
     }
 
     //Create checkboxes for each event
@@ -131,61 +130,334 @@ import { makePostCall, addNameList, findPos, mapAssignee, mapFromElement} from "
             });
 
 
+
+
+
+
             //Bulk Control Buttons
-            // jq$("assigneeList").click(function(){
-
-            // });
-
-            // jq$("wfOwnerList").click(function(){
-
-            // });
-
-            jq$("#moveAllRight").click(function(){
+            // 0 : Assignee List
+            // 1 : Workflow Owner List
+            var status = [true, false];
+            jq$('#button-wrapper').click(function(e){
                 jq$(".list-user").remove();
                 jq$(".list-assigned").remove();
-                addNameList("assign");
 
+                var event = jq$(e.target);
+                if(event.is("#assigneeList")){
+                    jq$("#user-list .flex-first-row").text("Name");
+                    jq$("#assigned-list .flex-first-row").text("Assignee");
+                    status[0] = true
+                    status[1] = false
+                    addNameList("default", names.assignee);
+                }
+
+                else if(event.is("#wfOwnerList")){
+                    jq$("#user-list .flex-first-row").text("Workflow Owner");
+                    jq$("#assigned-list .flex-first-row").text("Event");
+                    status[0] = false
+                    status[1] = true
+                    addNameList("default", names.wf_owner);
+                }
             });
 
-            jq$("#moveAllLeft").click(function(){
+
+
+            jq$('#moveAllLeft').click(function(){
                 jq$(".list-user").remove();
                 jq$(".list-assigned").remove();
-                addNameList("default");
+                if(status[0] == true){
+                    addNameList("default", names.assignee);
+                }
+                else if(status[1] == true){
+                    addNameList("default", names.wf_owner);
+                }
             });
+
+            jq$('#moveAllRight').click(function(){
+                jq$(".list-user").remove();
+                jq$(".list-assigned").remove();
+                if(status[0] == true){
+                    addNameList("assign", names.assignee);
+                }
+                else if(status[1] == true){
+                    addNameList("assign", names.wf_owner);
+                }
+            });
+
 
             jq$("#applySubmit").click(function(){
                 var finalized_name = mapFromElement(jq$('.list-assigned'));
-                makePostCall('addTaskAssignee', finalized_name, tableID, selected_office, selected_cu);
-                var save_data = {
-                    method: "update",
-                    programOfficeId: selected_office,
-                    displayPage: false,
-                    costingUnitId: selected_cu,
-                    processIndex: 0,
-                    taskIndex: 0,
-                    roleId: 0,
-                    taskAssigneeId: 0,
-                };
-
-                //Save data after adding users
-                jq$.ajax({
-                    type: "POST",
-                    async: false,
-                    data: save_data,
-                    url: 'https://'
-                    + ((d1$.isEmpty(d1$.hostname)) ? window.location.hostname : d1$.hostname)
-                    +'/srs/sysadmin/system/programOfficeTaskSetup.do?',
-
-                    success : function(data) {
-                        if(debugging) console.log('Data: '+ data);
-                    },
-                    error : function(msg)
-                    {
-                        if(debugging) console.log("Update FAILED: "+ msg);
-                    }
-                });
+                if(status[0] == true){
+                    postAssignee(finalized_name, tableID, selected_office, selected_cu);
+                }
+                else if(status[1] == true){
+                    //TODO: collect overseer events, change table view when clicked on button
+                    postOverseer(finalized_name, tableID, selected_office, selected_cu);
+                }
                 window.location.reload();
             });
 
+
+
+
+
+
+
+
+
+
+    /*********Functions**********/
+    function postAssignee(nameList, eventID, program_office, costing_unit){
+        if(isNaN(costing_unit)){
+            costing_unit = "";
+        }
+
+        //Static data
+        var postData = {
+            method: 'addTaskAssignee',
+            programOfficeId: program_office,
+            displayPage: false,
+            costingUnitId: costing_unit,
+            roleId: 0,
+        };
+
+        var saveData = {
+            method: "update",
+            programOfficeId: selected_office,
+            displayPage: false,
+            costingUnitId: selected_cu,
+            processIndex: 0,
+            taskIndex: 0,
+            roleId: 0,
+            taskAssigneeId: 0,
+        };
+
+        //Dynamic data
+        var eventIndex;
+        jq$.each(nameList, function(name, id){
+            for (var key = 0; key < eventID.length; key++) {
+                postData['processIndex'] = key
+                for(var value = 0; value < eventID[key].length; value++){
+                    eventIndex = eventID[key][value];
+                    postData['taskIndex'] = eventIndex
+                    postData['taskAssigneeId'] = id;
+                    postData['taskAssigneesInfo[' + key + '].selectedRole[' + eventIndex + ']'] = "";
+                    postData['taskAssigneesInfo[' + key + '].selectedAssignee[' + eventIndex + ']'] = id;
+                    postData['taskAssigneesInfo[' + key + '].ccMail[' + eventIndex + ']'] = "";
+
+                    jq$.ajax({
+                        type: "POST",
+                        async: false,
+                        data: postData,
+                        url: 'https://'
+                        + ((d1$.isEmpty(d1$.hostname)) ? window.location.hostname : d1$.hostname)
+                        +'/srs/sysadmin/system/programOfficeTaskSetup.do?',
+
+                        success : function(data) {
+                            if(debugging) console.log('Data: '+ data);
+                        },
+                        error : function(msg)
+                        {
+                            if(debugging) console.log("Update FAILED: "+ msg);
+                        }
+                    });
+                }
+            }
+        });
+
+        //Save data after adding users
+        jq$.ajax({
+            type: "POST",
+            async: false,
+            data: saveData,
+            url: 'https://'
+            + ((d1$.isEmpty(d1$.hostname)) ? window.location.hostname : d1$.hostname)
+            +'/srs/sysadmin/system/programOfficeTaskSetup.do?',
+
+            success : function(data) {
+                if(debugging) console.log('Data: '+ data);
+            },
+            error : function(msg)
+            {
+                if(debugging) console.log("Update FAILED: "+ msg);
+            }
+        });
+    }
+
+    function postOverseer(overseer_name, eventID, program_office, costing_unit){
+        if(isNaN(costing_unit)){
+            costing_unit = "";
+        }
+
+        //Static data
+        var postData = {
+            method: 'update',
+            programOfficeId: program_office,
+            displayPage: false,
+            costingUnitId: costing_unit,
+            roleId: 0,
+        };
+
+        //Only 1 name for process overseer
+        for (var key = 0; key < eventID.length; key++) {
+            postData['processOverseer[' + key + ']'] = overseer_name;
+        }
+
+        jq$.ajax({
+            type: "POST",
+            async: false,
+            data: postData,
+            url: 'https://'
+            + ((d1$.isEmpty(d1$.hostname)) ? window.location.hostname : d1$.hostname)
+            +'/srs/sysadmin/system/programOfficeTaskSetup.do?',
+
+            success : function(data) {
+                if(debugging) console.log('Data: '+ data);
+            },
+            error : function(msg)
+            {
+                if(debugging) console.log("Update FAILED: "+ msg);
+            }
+        });
+    }
+
+
+    //default = push all names to left of table
+    //assign = push all names to right of table
+    function addNameList(mode, nameList){
+        if(mode == "default" && nameList == names.assignee){
+            jq$.each(nameList, function(name, id){
+                jq$('<li>', {
+                class: 'flex-list-items list-user',
+                value: id,
+                text: name,
+                }).appendTo('#user-list');
+            });
+            jq$.each(nameList, function(){
+                jq$('<li>', {
+                class: 'flex-list-items list-assigned',
+                }).appendTo('#assigned-list');
+            });
+        }
+
+        else if(mode == "assign" && nameList == names.assignee){
+                jq$.each(nameList, function(){
+                    jq$('<li>', {
+                    class: 'flex-list-items list-user',
+                    }).appendTo('#user-list');
+                });
+                jq$.each(nameList, function(name, id){
+                    jq$('<li>', {
+                    class: 'flex-list-items list-assigned',
+                    value: id,
+                    text: name,
+                    }).appendTo('#assigned-list');
+            });
+        }
+
+        else if(mode == "default" && nameList == names.wf_owner){
+            jq$.each(nameList, function(name, id){
+                jq$('<li>', {
+                class: 'flex-list-items list-user',
+                value: id,
+                text: name,
+                }).appendTo('#user-list');
+            });
+            jq$.each(nameList, function(){
+                jq$('<li>', {
+                class: 'flex-list-items list-assigned',
+                }).appendTo('#assigned-list');
+            });
+    }
+
+        else if(mode == "assign" && nameList == names.wf_owner){
+            jq$.each(nameList, function(){
+                jq$('<li>', {
+                class: 'flex-list-items list-user',
+                }).appendTo('#user-list');
+            });
+            jq$.each(nameList, function(name, id){
+                jq$('<li>', {
+                class: 'flex-list-items list-assigned',
+                value: id,
+                text: name,
+                }).appendTo('#assigned-list');
+        });
+    }
+    }
+
+
+    //Find index of keys from dictionary
+    function findPos(dict, target){
+        var index = 1;
+        for(var name in dict){
+            if(name == target){
+                break;
+            }
+            ++index;
+        }
+        return index;
+    }
+
+    //Extract names from dropdown box
+    function mapAssignee(program_office, costing_unit){
+        var assignee_selector = null;
+        var wf_owner_selector = null;
+
+        if(program_office == 5){
+            assignee_selector = jq$('[name="taskAssigneesInfo[0].selectedAssignee[2]"] option');
+        }
+
+        //These program office have the same html tag names
+        else if(program_office == 18153 || program_office == 18356){
+            //value of first costing unit if exist is NaN
+            if(isNaN(costing_unit)){
+                assignee_selector = jq$('[name="taskAssigneesInfo[16].selectedAssignee[1]"] option');
+                wf_owner_selector = jq$('[name="processOverseer[16]"] option');
+            }
+            else{
+                assignee_selector = jq$('[name="taskAssigneesInfo[0].selectedAssignee[1]"] option');
+                wf_owner_selector = jq$('[name="processOverseer[0]"] option');
+            }
+        }
+
+        else if(program_office == 4){
+            assignee_selector = jq$('[name="costingUnitDirectors[0]"] option');
+        }
+
+        else if(program_office == 3){
+            assignee_selector = jq$('[name="taskAssigneesInfo[0].selectedAssignee[1]"] option');
+            wf_owner_selector = jq$('[name="processOverseer[0]"] option');
+        }
+
+        else if(program_office == 2){
+            assignee_selector = jq$('[name="taskAssigneesInfo[1].selectedAssignee[1]"] option');
+            wf_owner_selector = jq$('[name="processOverseer[2]"] option');
+        }
+
+        else if(program_office == 1){
+            assignee_selector = jq$('[name="taskAssigneesInfo[50].selectedAssignee[1]"] option');
+            wf_owner_selector = jq$('[name="processOverseer[50]"] option');
+        }
+
+        var assignee = mapFromElement(assignee_selector);
+        var wf_owner = mapFromElement(wf_owner_selector);
+
+
+        return {"assignee": assignee, "wf_owner" : wf_owner};
+    }
+
+
+    //Convert text and value to a list
+    function mapFromElement(HTML_selector){
+        var result = {};
+
+        jq$(HTML_selector).each(function(){
+            if(jq$(this).text() != ""){
+                result[jq$(this).text()] = jq$(this).val();
+            }
+        })
+        return result;
+    }
 })(jQuery);
